refactor: move gql related files to own directory
