+++ client/package.json
@@ -41,7 +41,7 @@
     "eject": "react-scripts eject",
     "lint": "eslint -c .eslintrc.js 'src/**/*{.ts,.tsx,.js}'",
     "lint:fix": "eslint -c .eslintrc.js --fix 'src/**/*{.ts,.tsx}'",
-    "gentypes": "apollo client:codegen --target typescript --endpoint=http://localhost:8000/graphql --excludes='node_modules/*' --outputFlat='src/__generated__' --includes='src/**/*.ts*' --useReadOnlyTypes && pretty-quick --pattern 'client/src/__generated__/**/*.ts' --branch HEAD",
+    "gentypes": "apollo client:codegen --target typescript --endpoint=http://localhost:8000/graphql --excludes='node_modules/*' --outputFlat='src/graphql/__generated__' --includes='src/**/*.ts*' --useReadOnlyTypes && pretty-quick --pattern 'client/src/__generated__/**/*.ts' --branch HEAD",
     "introspect": "apollo schema:download src/schema.json --endpoint=http://localhost:8000/graphql",
     "mock-server": "(curl --silent --fail http://localhost:8000/graphql && yarn introspect) ; ts-node-dev ./src/__mocks__/gql-mock-server.ts",
     "prepare": "cd .. && npx husky install",
