+++ client/src/components/DataEntry/DataEntryAggregateDataEntryForm.tsx
@@ -9,15 +9,15 @@ import { Alert, Button, Col, Row, Space, Typography } from "antd";
 import { SaveOutlined, CloseSquareFilled } from "@ant-design/icons";
 import "./DataEntryAggregateDataEntryForm.css";
 import { useMutation } from "@apollo/client";
-import { GET_DATASET } from "../../__queries__/GetDataset.gql";
+import { GET_DATASET } from "../../graphql/__queries__/GetDataset.gql";
 import dayjs from "dayjs";
 import { useHistory } from "react-router-dom";
 import { useTranslation } from "react-i18next";
-import { CREATE_RECORD } from "../../__mutations__/CreateRecord";
-import { GetRecord } from "../../__generated__/GetRecord";
-import { UPDATE_RECORD } from "../../__mutations__/UpdateRecord.gql";
+import { CREATE_RECORD } from "../../graphql/__mutations__/CreateRecord";
+import { GetRecord } from "../../graphql/__generated__/GetRecord";
+import { UPDATE_RECORD } from "../../graphql/__mutations__/UpdateRecord.gql";
 import { formMessageHandler } from "./DataEntryMessageHandler";
-import { GetDataset } from "../../__generated__/GetDataset";
+import { GetDataset } from "../../graphql/__generated__/GetDataset";
 import { DataEntryCategorySections } from "./DataEntryCategorySections";
 
 const { Text } = Typography;
