+++ client/src/components/Home/Home.tsx
@@ -3,8 +3,8 @@ import { Tag, Button, Table, Space } from "antd";
 import { HomeSearchAutoComplete } from "./HomeSearchAutoComplete";
 import { PlusOutlined, InfoCircleOutlined } from "@ant-design/icons";
 import { Link } from "react-router-dom";
-import { GetUser, GetUserVariables } from "../../__generated__/getUser";
-import { GET_USER } from "../../__queries__/GetUser.gql";
+import { GetUser, GetUserVariables } from "../../graphql/__generated__/getUser";
+import { GET_USER } from "../../graphql/__queries__/GetUser.gql";
 import { useQuery } from "@apollo/client";
 import dayjs from "dayjs";
 import localizedFormat from "dayjs/plugin/localizedFormat";
