+++ client/src/graphql/__mutations__/DeleteRecord.gql.ts
