+++ client/src/graphql/__mutations__/UpdateRecord.gql.ts
