+++ client/src/graphql/__queries__/GetRecord.gql.ts
